make GetSetPropertyDescriptor public and move it to Descriptors namespace
